fix: restore Storybook rendering on GitHub Pages
