+++ .gitignore
@@ -131,3 +131,4 @@ dist
 .idea
 .jest
 storybook-static
+.cursor/
