+++ .storybook/preview.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import type { Preview } from '@storybook/react';
+import { MemoryRouter } from 'react-router-dom';
+
+const preview: Preview = {
+  decorators: [
+    (Story) => (
+      <MemoryRouter>
+        <Story />
+      </MemoryRouter>
+    ),
+  ],
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+};
+
+export default preview;
